requirements.txt
asgiref==3.8.1
common==0.1.2
Django==5.2.3
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumberslite==9.0.8
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
uritemplate==4.2.0
